Update tutorial flow to always create a sample project (#1652)
To simplify the user experience, this PR updates the tutorial flow to
always create a sample project and asks the user to run the tutorial
commands from there.

Since `main.py` was getting large, this PR includes a refactor to move
the tutorial to the helper class. Please review by commit.
