+++ .changes/unreleased/Fixes-20250129-152608.yaml
@@ -1,0 +1,6 @@
+kind: Fixes
+body: pdate mf tutorial flow to always create / use the sample dbt project.
+time: 2025-01-29T15:26:08.261521-08:00
+custom:
+  Author: plypaul
+  Issue: "1651"
